quickfix on "permalink to revision" issue
http://wordpress.org/support/topic/404-error-potentially-caused-by-revisions

In some cases "custom permalink" creates TWO permalinks with same url:
one for a post/page and another one on this post/page's revision. This fix 
avoids saving of permalinks on revisions and maybe on something else.

Works perfect for me, but didn't tested well on all cases. so may corrupt 
something else :")